Refactor Matter IDs to typed value classes (UInt/ULong)

+++ app/src/main/java/io/aether/android/chip/SubscriptionHelper.kt
@@ -79,9 +79,9 @@ class SubscriptionHelper @Inject constructor(private val chipClient: ChipClient)
     nodeState.endpointStates.forEach { (_endpointId, endpointState) ->
       if (_endpointId != endpointId) return@forEach
       endpointState.clusterStates.forEach { (clusterId, clusterState) ->
-        if (clusterId != clusterAttribute.clusterId) return@forEach
+        if (clusterId != clusterAttribute.clusterId.toLong()) return@forEach
         clusterState.attributeStates.forEach { (attributeId, attributeState) ->
-          if (attributeId != clusterAttribute.attributeId) return@forEach
+          if (attributeId != clusterAttribute.attributeId.toLong()) return@forEach
           return attributeState.value
         }
       }
